listen logs view in Logs folder, and send mqtt message in new folder
